requirements.txt
appnope==0.1.0
asgiref==3.2.10
autopep8==1.5.3
backcall==0.2.0
certifi==2020.6.20
chardet==3.0.4
decorator==4.4.2
Django==3.0.8
django-extensions==3.0.2
djangorestframework==3.11.0
idna==2.10
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.17.1
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.5
ptyprocess==0.6.0
pycodestyle==2.6.0
Pygments==2.6.1
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
traitlets==4.3.3
urllib3==1.25.9
wcwidth==0.2.5
